+++ helpers/functions.py
@@ -1,13 +1,14 @@
+"""Assets Browser Functions"""
 import os
 import sys
-import ctypes
 import logging
 import platform
 import subprocess
 
 from PyQt5 import QtGui, QtCore, QtWidgets
 from config import configurations, constants
-from modules.utils import alert_window
+from helpers.exceptions import InvalidProjectPath
+from helpers.utils import alert_window
 
 logger = logging.getLogger(__name__)
 
@@ -235,9 +236,9 @@ def column_views(column_view, category, project):
 
 
 def project_list(self):
-    """List Project directories in PROJECTPATH comboBox.
+    """List Project directories in PROJECT_PATH comboBox.
 
-    Retrieve directories in PROJECTPATH comboBox, clear existing tabs and create new tabs.
+    Retrieve directories in PROJECT_PATH comboBox, clear existing tabs and create new tabs.
 
     """
     # 1. Update INI CurrentProject with chosen project from comboBox
@@ -279,16 +280,19 @@ def valid_path(ini, project):
     ini : str
         Path to INI file.
     project : str
-        Project name.
+        Path to project directory.
 
     Returns
     -------
-    bool
-        True if exists, else False.
+    None
+
+    Raises
+    ------
+    InvalidProjectPath
+        If project path value in INI is invalid.
 
     """
     exists = os.path.exists(project)
-
     if not exists:
         # 1. Set Project Path to User's Home directory
         home = os.path.expanduser('~')
@@ -316,8 +320,7 @@ def valid_path(ini, project):
         )
 
         alert_window('Warning', warning_text)
-
-    return exists
+        raise InvalidProjectPath(project)
 
 
 def clear_layout(layout):
@@ -338,24 +341,6 @@ def show_debug(self):
     else:
         text.setHidden(True)
         text.setEnabled(False)
-
-
-def always_on_top(self):
-    """Toggle AlwaysOnTop (works in Windows and Linux)."""
-    if self.actionAlwaysOnTop.isChecked():
-        self.setWindowFlags(self.windowFlags() | QtCore.Qt.WindowStaysOnTopHint)
-        print("Always on Top Enabled")
-    else:
-        self.setWindowFlags(self.windowFlags() & ~QtCore.Qt.WindowStaysOnTopHint)
-        print("Always on Top Disabled")
-    self.show()
-
-
-def center_screen(self):
-    """Center PyQt Window on screen."""
-    resolution = QtWidgets.QDesktopWidget().screenGeometry()
-    self.move((resolution.width() / 2) - (self.frameSize().width() / 2),
-              (resolution.height() / 2) - (self.frameSize().height() / 2))
 
 
 def get_file_size(size, precision=2):
@@ -421,58 +406,6 @@ def reveal_in_os(path):
     subprocess.call(cmd)
 
 
-def font_size_overrides(self, size=8, scale=1.0):
-    """Overrides font sizes.
-
-    Overrides font sizes based on platform due to PyQt quirks especially on macOS/OSX.
-
-    Notes
-    -----
-    This can also override the default font size to arbitrary values although the default
-    values are good enough on non HiDPI display (e.g. Windows 7).
-
-    Parameters
-    ----------
-    size : int
-        Set the default font size.
-    scale : float
-        The scale multiplier to resize the font size.
-
-    Returns
-    -------
-    None
-
-    """
-    system = platform.system()
-    font = QtGui.QFont()
-
-    if system == 'Darwin':
-        size = 12
-        scale = 1.0
-    if system == 'Linux':
-        scale = 1.0
-
-    font_size = size
-    font.setPointSize(font_size * scale)
-    self.setFont(font)
-
-
-def set_window_icon(self, icon='icons/file.png'):
-    """Set PyQt Window Icon.
-
-    Parameters
-    ----------
-    icon : str
-        Path to icon file.
-
-    Returns
-    -------
-    None
-
-    """
-    self.setWindowIcon(QtGui.QIcon(icon))
-
-
 def open_file(target):
     """ Open selected file using the OS associated program.
 
@@ -496,34 +429,6 @@ def open_file(target):
             subprocess.call(['open', target])
 
 
-def hidpi_check(app):
-    """HiDPI check for QApplication.
-
-    Parameters
-    ----------
-    app : PyQt5.QtWidgets.QApplication
-        PyQt QtWidgets QApplication object.
-
-    Returns
-    -------
-    None
-
-    """
-    if hasattr(QtCore.Qt, 'AA_EnableHighDpiScaling'):
-        app.setAttribute(QtCore.Qt.AA_EnableHighDpiScaling, True)
-        logger.info('High DPI Scaling Enabled')
-    if hasattr(QtCore.Qt, 'AA_UseHighDpiPixmaps'):
-        app.setAttribute(QtCore.Qt.AA_UseHighDpiPixmaps, True)
-        logger.info('High DPI Pixmaps Enabled')
-
-
-def taskbar_icon():
-    """Workaround to show setWindowIcon on Win7 taskbar instead of default Python icon."""
-    if platform.system() == 'Windows':
-        app_id = u'taukeke.python.assetsbrowser'  # arbitrary string
-        ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(app_id)
-
-
 def show_cwd():
     """Show CWD (Current Work Directory) as a QMessageBox."""
     widget = QtWidgets.QWidget()
@@ -541,10 +446,6 @@ def restart_app():
 
     99% it doesn't restart in an IDE like PyCharm for complex script but
     it has been tested to work when execute through Python interpreter.
-
-    Returns
-    -------
-    None
 
     """
     os.execv(sys.executable, [sys.executable] + sys.argv)