+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 ## Unreleased
 ### Core
 - Dynamically generate radio buttons in `Create New Asset` dialog from `CategoryList` in INI file.
-- Minor update to `AssetDialog` UI.
 - Add CHANGELOG.
+
+### UI
+- Minor update to `AssetDialog` UI.
+- Update Preferences Dialog UI.
+- Add `Settings`, `Assets`, and `Advanced` pages in Preferences Dialog UI.
 
 ## 0.0.1 (2018-11-19)
 ### Core